Chore - Restructure codebase
