+++ clypi/__init__.py
@@ -1,11 +1,17 @@
 from clypi import parsers
 from clypi._align import AlignType, align
-from clypi._arg_config import Positional, arg
-from clypi._boxed import Boxes, boxed
-from clypi._cli import ClypiFormatter, Command, Formatter
+from clypi._cli.arg_config import Positional, arg
+from clypi._cli.distance import closest, distance
+from clypi._cli.formatter import ClypiFormatter, Formatter
+from clypi._cli.main import Command
 from clypi._colors import ALL_COLORS, ColorType, Styler, cprint, style
+from clypi._components.boxed import Boxes, boxed
+from clypi._components.indented import indented
+from clypi._components.separator import separator
+from clypi._components.spinners import Spin, Spinner, spinner
+from clypi._components.stack import stack
+from clypi._components.wraps import OverflowStyle, wrap
 from clypi._configuration import ClypiConfig, Theme, configure, get_config
-from clypi._distance import closest, distance
 from clypi._exceptions import (
     AbortException,
     ClypiException,
@@ -13,15 +19,10 @@ from clypi._exceptions import (
     format_traceback,
     print_traceback,
 )
-from clypi._indented import indented
 from clypi._prompts import (
     confirm,
     prompt,
 )
-from clypi._separator import separator
-from clypi._spinners import Spin, Spinner, spinner
-from clypi._stack import stack
-from clypi._wraps import OverflowStyle, wrap
 from clypi.parsers import Parser
 
 __all__ = (
