+++ clypi/_cli/autocomplete.py
@@ -8,10 +8,11 @@ from abc import ABC, abstractmethod
 from pathlib import Path
 from textwrap import dedent
 
-from clypi import _arg_parser, _colors
+from clypi import _colors
+from clypi._cli import arg_parser
 
 if t.TYPE_CHECKING:
-    from clypi._cli import Command
+    from clypi._cli.main import Command
 
 _CLYPI_CURRENT_ARGS = "_CLYPI_CURRENT_ARGS"
 
@@ -170,7 +171,7 @@ def list_arguments(command: type[Command]):
 def requested_autocomplete_install(args: t.Sequence[str]) -> bool:
     if not args:
         return False
-    parsed = _arg_parser.parse_as_attr(args[-1])
+    parsed = arg_parser.parse_as_attr(args[-1])
     return parsed.is_long_opt() and parsed.value == "install_autocomplete"
