+++ clypi/_components/boxed.py
@@ -3,9 +3,9 @@ import typing as t
 from clypi._align import AlignType
 from clypi._align import align as _align
 from clypi._colors import ColorType, Styler
+from clypi._components.wraps import wrap
 from clypi._data.boxes import Boxes as _Boxes
 from clypi._util import get_term_width, visible_width
-from clypi._wraps import wrap
 
 Boxes = _Boxes
