--- clypi/_const.py
@@ -1,1 +1,0 @@
-ESC = "\033["
